Add ContentEncodingTrait  to remove code duplication
